third_party/WebKit/LayoutTests/fast/dom/Window/window-scroll-arguments-expected.txt

This test makes sure that calling the window scrolling methods with less than 2 arguments throws exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



window.scrollTo Tests

Testing - scrollTo with 0 arguments
PASS window.scrollTo() threw exception TypeError: Failed to execute 'scrollTo' on 'Window': 2 arguments required, but only 0 present..
Testing - scrollTo with 1 argument
PASS window.scrollTo(x) threw exception TypeError: Failed to execute 'scrollTo' on 'Window': 2 arguments required, but only 1 present..
Testing - scrollTo with more than 2 arguments
PASS window.scrollX is x
PASS window.scrollY is y

window.scroll Tests

Testing - scroll with 0 arguments
PASS window.scroll() threw exception TypeError: Failed to execute 'scroll' on 'Window': 2 arguments required, but only 0 present..
Testing - scroll with 1 argument
PASS window.scroll(x) threw exception TypeError: Failed to execute 'scroll' on 'Window': 2 arguments required, but only 1 present..
Testing - scroll with more than 2 arguments
PASS window.scrollX is x
PASS window.scrollY is y

window.scrollBy Tests

Testing - scrollBy with 0 arguments
PASS window.scrollBy() threw exception TypeError: Failed to execute 'scrollBy' on 'Window': 2 arguments required, but only 0 present..
Testing - scrollBy with 1 argument
PASS window.scrollBy(x) threw exception TypeError: Failed to execute 'scrollBy' on 'Window': 2 arguments required, but only 1 present..
Testing - scrollBy with more than 2 arguments
PASS window.scrollX is resetX + x
PASS window.scrollY is resetY + y
PASS successfullyParsed is true

TEST COMPLETE
This box should force the window to have a scrollable area to test.
